refactor(components): Refactor export statements
Deleted component folders
Refactored named exports to default exports in the following files:
- src/components/Wrapper.tsx
- src/components/Header.tsx
- src/pages/Login.tsx
- src/pages/Register.tsx
- src/pages/index.ts
