+++ src/components/Wrapper.tsx
@@ -1,7 +1,8 @@
 import React from "react";
 import styled from "styled-components";
-import { Header } from "components/Header/Header";
+
 import { colors } from "shared/styles";
+import { Header } from "components";
 
 const Container = styled.div`
   background-color: ${colors.primary};
@@ -21,4 +22,4 @@ const Wrapper = ({ children }: WrapperProps) => {
   );
 };
 
-export { Wrapper };
+export default Wrapper;
