+++ src/pages/Login.tsx
@@ -4,4 +4,4 @@ const Login = () => {
   return <h1>Login Page</h1>;
 };
 
-export { Login };
+export default Login;
