+++ src/pages/Register.tsx
@@ -4,4 +4,4 @@ const Register = () => {
   return <h1>Register Page</h1>;
 };
 
-export { Register };
+export default Register;
