+++ src/pages/index.ts
@@ -1,2 +1,7 @@
-export { Login } from "./Login";
-export { Register } from "./Register";
+import Login from './Login'
+import Register from './Register'
+
+export {
+    Login,
+    Register
+}
